refactor(actor-sheet): #83 phase 2 — extract roll + inventory-transfer helpers
Pulls the four roll dispatchers and two cross-actor inventory ops out of
actor-sheet.ts into dedicated helpers:

  - sheet-helpers/rolls.ts:
      rollAvailableVehicleAction, rollAvailableAction,
      rollBodyPoints, rollPurchase
  - sheet-helpers/inventory-transfer.ts:
      onPurchase, onTransfer

All six methods on OD6SActorSheet keep their public names (each has at
least one caller in another module: combat-actions / scores listeners,
chat-log-listeners, handle-wild-die, roll-execute, socketlib's purchase
finalize, etc.) and now delegate one-line into the helpers — same
pattern as the existing deleteItem/addItem/_onDrop* delegates.

actor-sheet.ts: 651 -> 473 LOC. Phase 3 (drag/crew consolidation) brings
it under the issue's <~400 LOC acceptance; phase 4 (ActorSheetContext
typing) remains gated on #57 closing the augmentation gaps.

No new unit tests — these methods are Foundry-coupled (Roll, ChatMessage,
od6sroll._onRollDialog, ui.notifications). Coverage from existing smoke
specs: tier-3-roll, tier-3-skill-roll, tier-3-weapon-roll,
tier-3-action-rolls, tier-3-vehicle, tier-2-sheets all green.
